requirements/base.txt
Django==1.8.8
django-environ>=0.4.0
django-braces>=1.8.1
django-crispy-forms>=1.5.2
django-admin-bootstrapped>=2.5.6
easy-thumbnails>=2.3
django-notifications-hq==0.8.0
beautifulsoup4==4.4.0
django-contrib-comments==1.6.1
django-mptt==0.7.4
django-tagging==0.4
django-xmlrpc==0.1.5
mots-vides==2015.5.11
pillow==2.9.0
pyparsing==2.0.3
